deploy/requirements.txt
psycopg2-binary==2.8.5
gunicorn
pandas
scikit-learn
dash
dash-bootstrap-components
dash-core-components
dash-html-components
dash-table
folium
